Controller e Service

+++ src/main/java/com/sistema/delivery/repository/EnderecoRepositpry.java
@@ -4,6 +4,6 @@ import com.sistema.delivery.domian.Endereco;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoRepositpry extends JpaRepository<Endereco, Long> {
+public interface EnderecoRepositpry extends JpaRepository<Endereco, Integer> {
     
 }
